Refactor roomGenerator architecture and add system documentation
- Add `docs/SYSTEM_MAP.md` detailing system architecture, API, and workflows.
- Refactor backend: Extract `Project` structs to `models.go` and default assets to `defaults.go`.
- Refactor frontend: Extract project loading logic to `src/domain/projectService.js`.
- Update `CONTRIBUTING.md` to reference new documentation and structure.
